Create SQL to join tables
In creating a query to find the top 5 customers from the top 10 cities who have paid the highest total amounts to Rockbuster the tip structure was used. I used SELECT for SELECT; A.customer_id, A.first_name, A.last_name, D.country, C.city, and SUM for (B.amount) AS total_amount_paid. I also included the LIMIT to 5 so the query could only return 5 top most customers in the 10 cities as seen in the output above.